+++ CHANGES.rst
@@ -34,9 +34,14 @@ Unreleased
 
 - The new 3.12 soft keyword ``type`` is properly bolded in HTML reports.
 
+- Removed the "fullcoverage" feature used by CPython to measure the coverage of
+  early-imported standard library modules.  CPython `stopped using it
+  <88054_>`_ in 2021, and it stopped working completely in Python 3.13.
+
 .. _issue 1605: https://github.com/nedbat/coveragepy/pull/1605
 .. _issue 1684: https://github.com/nedbat/coveragepy/issues/1684
 .. _pull 1685: https://github.com/nedbat/coveragepy/pull/1685
+.. _88054: https://github.com/python/cpython/issues/88054
 
 
 .. scriv-start-here